+++ ModDebuggingMod/Card.cs
@@ -15,7 +15,7 @@ namespace ModDebuggingMod
         {
             CardInfo info = CardHelper.NewCard(true, "wstl", "wstlcard", "Debug",
                 attack: 1, health: 100, blood: 0, bones: 0, energy: 0, gems: null)
-                .AddAbilities(ParalysingStrike.ability)
+                .AddAbilities(FrostRuler.ability)
                 .AddSpecialAbilities(Bind.specialAbility)
                 //.SetTransformerCardId("Squirrel")//.SetEvolve("Squirrel", 1)
                 .SetPortraits("misterWin_grimora", emissionName: "misterWin_grimora_emission", pixelPortraitName: "buffBell.png")