+++ WhistleWind.AbnormalSigils/Abilities/abilityFalseThrone.cs
@@ -30,8 +30,13 @@ namespace WhistleWind.AbnormalSigils
         public override int StartingHealthCost => 1;
         public override int TurnDelay => 1;
 
-        public override bool CardSlotCanBeTargeted(CardSlot slot) => slot.Card != null && slot.Card != base.Card && slot.Card.LacksAllTraits(Trait.Giant, Trait.Uncuttable);
+        public override bool IsInvalidTarget(CardSlot slot)
+        {
+            if (slot.Card != null && slot.Card != base.Card)
+                return slot.Card.HasAnyOfTraits(Trait.Giant, Trait.Uncuttable);
 
+            return base.IsInvalidTarget(slot);
+        }
         public override IEnumerator OnValidTargetSelected(CardSlot slot)
         {
             if (slot != null && slot.Card != null)